chore(admin): convert useSettingsForm to TS (#18608)
