+++ packages/core/admin/admin/src/hooks/index.js
@@ -1,2 +1,1 @@
-export { default as useSettingsForm } from './useSettingsForm';
 export { default as useSettingsMenu } from './useSettingsMenu';
