--- packages/core/admin/admin/src/hooks/useSettingsForm/index.js
@@ -1,168 +1,0 @@
-import { useEffect, useReducer } from 'react';
-
-import {
-  getYupInnerErrors,
-  useFetchClient,
-  useNotification,
-  useOverlayBlocker,
-} from '@strapi/helper-plugin';
-import omit from 'lodash/omit';
-
-import { formatAPIErrors } from '../../utils/formatAPIErrors';
-
-import init from './init';
-import { initialState, reducer } from './reducer';
-
-const checkFormValidity = async (data, schema) => {
-  let errors = null;
-
-  try {
-    await schema.validate(data, { abortEarly: false });
-  } catch (err) {
-    errors = getYupInnerErrors(err);
-  }
-
-  return errors;
-};
-
-/**
- * TODO: refactor this, it's confusing and hard to read.
- * It's also only used in `Settings/pages/SingleSignOn` so it can
- * probably be deleted and everything written there...
- */
-const useSettingsForm = (endPoint, schema, cbSuccess, fieldsToPick) => {
-  const [
-    { formErrors, initialData, isLoading, modifiedData, showHeaderButtonLoader, showHeaderLoader },
-    dispatch,
-  ] = useReducer(reducer, initialState, () => init(initialState, fieldsToPick));
-  const toggleNotification = useNotification();
-  const { lockApp, unlockApp } = useOverlayBlocker();
-
-  const { get, put } = useFetchClient();
-
-  useEffect(() => {
-    const getData = async () => {
-      try {
-        const {
-          data: { data },
-        } = await get(endPoint);
-
-        dispatch({
-          type: 'GET_DATA_SUCCEEDED',
-          data,
-          fieldsToPick,
-        });
-      } catch (err) {
-        console.error(err.response);
-        toggleNotification({
-          type: 'warning',
-          message: { id: 'notification.error' },
-        });
-      }
-    };
-
-    if (endPoint) {
-      getData();
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [endPoint]);
-
-  const handleCancel = () => {
-    dispatch({
-      type: 'ON_CANCEL',
-    });
-  };
-
-  const handleChange = ({ target: { name, value, type: inputType } }) => {
-    dispatch({
-      type: 'ON_CHANGE',
-      inputType,
-      keys: name,
-      value,
-    });
-  };
-
-  const setField = (fieldName, value) => {
-    dispatch({
-      type: 'ON_CHANGE',
-      keys: fieldName,
-      value,
-    });
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    const errors = await checkFormValidity(modifiedData, schema);
-
-    dispatch({
-      type: 'SET_ERRORS',
-      errors: errors || {},
-    });
-
-    if (!errors) {
-      try {
-        lockApp();
-
-        dispatch({
-          type: 'ON_SUBMIT',
-        });
-        const cleanedData = omit(modifiedData, ['confirmPassword', 'registrationToken']);
-
-        if (cleanedData.roles) {
-          cleanedData.roles = cleanedData.roles.map((role) => role.id);
-        }
-        if (cleanedData.ssoLockedRoles) {
-          cleanedData.ssoLockedRoles = [...new Set(cleanedData.ssoLockedRoles)];
-        }
-
-        const {
-          data: { data },
-        } = await put(endPoint, cleanedData);
-
-        cbSuccess(data);
-
-        dispatch({
-          type: 'ON_SUBMIT_SUCCEEDED',
-          data,
-        });
-
-        toggleNotification({
-          type: 'success',
-          message: { id: 'notification.success.saved' },
-        });
-      } catch (err) {
-        const data = err?.response?.payload ?? { data: {} };
-
-        if (!!data?.data && typeof data.data === 'string') {
-          toggleNotification({
-            type: 'warning',
-            message: data.data,
-          });
-        } else {
-          toggleNotification({
-            type: 'warning',
-            message: data.message,
-          });
-        }
-
-        const apiErrors = formatAPIErrors(data);
-
-        dispatch({
-          type: 'SET_ERRORS',
-          errors: apiErrors,
-        });
-      } finally {
-        unlockApp();
-      }
-    }
-  };
-
-  return [
-    { formErrors, initialData, isLoading, modifiedData, showHeaderButtonLoader, showHeaderLoader },
-    dispatch,
-    { handleCancel, handleChange, handleSubmit, setField },
-  ];
-};
-
-export default useSettingsForm;
