--- packages/core/admin/admin/src/hooks/useSettingsForm/init.js
@@ -1,5 +1,0 @@
-const init = (initialState, fieldsToPick) => {
-  return { ...initialState, fieldsToPick };
-};
-
-export default init;
